long int and signedness fix from Nelson Beebe

+++ tls.c
@@ -115,9 +115,9 @@ static void list_file (const char *fname)
 	
 	/* NB: need to pass size as a double because it might be be
 	 * too large for a long. */
-	printf("%s %12.0f %6d.%-6d %6d %s %s%s\n",
+	printf("%s %12.0f %6ld.%-6ld %6d %s %s%s\n",
 	       permbuf, (double) buf.st_size,
-	       buf.st_uid, buf.st_gid,
+	       (long) buf.st_uid, (long) buf.st_gid,
 	       buf.st_nlink,
 	       datebuf, fname, linkbuf);
 }
